Feature/meeting

See merge request s07-webmobile1-sub2/S07P12A701!50

+++ BE/utf/src/main/java/com/ssafy/utf/api/request/RecordingReq.java
@@ -14,7 +14,7 @@ public class RecordingReq {
     private String name;
     private String sessionId;
     private long createdAt;
-    private float duration;
+    private long duration;
     private String url;
     private String chatRecord;
